moved policy models and added helper api methods to policy component
