+++ server/frontend/src/Routes.ts
@@ -6,24 +6,22 @@ const _returnRoute = (route: string) => {
     return route;
 }
 
-export default class Routes {
-    static requestHistoryRoutes = () => {
-        return {
-            getTransactionsRoute: _returnRoute("/request-history/transactions"),
-            getTransactionDetailsRoute: _returnRoute("/request-history/transactionDetails")
-        }
-    }
+const Routes = {
+    requestHistoryRoutes: {
+        getTransactionsRoute: _returnRoute("/request-history/transactions"),
+        getTransactionDetailsRoute: _returnRoute("/request-history/transactionDetails")
+    },
 
-    static policyRoutes = () => {
-        return {
-            getPolicyByIdRoute: _returnRoute("/policy/getPolicy"),
-            deletePolicyRoute: _returnRoute("/policy/deletePolicy"),
-            getDraftPolicyRoute: _returnRoute("/policy/getDraftPolicy"),
-            updateDraftPolicyRoute: _returnRoute("/policy/updateDraftPolicy"),
-            getCurrentPolicyRoute: _returnRoute("/policy/getCurrentPolicy"),
-            getPolicyHistory: _returnRoute("/policy/getPolicyHistory"),
-            publishPolicyRoute: _returnRoute("/policy/publishPolicy"),
-            distributePolicyRoute: _returnRoute("/policy/distributePolicy")
-        }
+    policyRoutes: {
+        getPolicyByIdRoute: _returnRoute("/policy/getPolicy"),
+        deletePolicyRoute: _returnRoute("/policy/deletePolicy"),
+        getDraftPolicyRoute: _returnRoute("/policy/getDraftPolicy"),
+        updateDraftPolicyRoute: _returnRoute("/policy/updateDraftPolicy"),
+        getCurrentPolicyRoute: _returnRoute("/policy/getCurrentPolicy"),
+        getPolicyHistory: _returnRoute("/policy/getPolicyHistory"),
+        publishPolicyRoute: _returnRoute("/policy/publishPolicy"),
+        distributePolicyRoute: _returnRoute("/policy/distributePolicy")
     }
 }
+
+export default Routes;
