+++ server/frontend/src/views/Policy/api/helpers/index.ts
@@ -1,0 +1,3 @@
+export { getPolicy } from "./getPolicy";
+
+export { getPolicyById } from "./getPolicyById";
