+++ server/frontend/src/views/Policy/api/index.ts
@@ -1,0 +1,6 @@
+import { getPolicy, getPolicyById } from "./helpers";
+import Routes from "../../../Routes";
+
+export const getCurrentPolicy = async (): Promise<string> => {
+    return await getPolicy(Routes.policyRoutes.getCurrentPolicyRoute);
+}
