+++ server/frontend/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "include": [
-    "src/**/*", "../src/common/models/**/*"
+    "src/**/*", "../src/common/models/PolicyManagementService/Policy", "../src/common/models/**/*"
   ],
   "exclude": [
     "node_modules",
