+++ server/src/common/models/PolicyManagementService/Policy/AdaptionPolicy/ContentFlags/ExcelContentFlags.ts
@@ -1,8 +1,9 @@
-import { ContentManagementFlagAction } from "../../../../../../src/common/models/enums/ContentManagementFlagAction";
+import { ContentManagementFlagAction } from "../../../../enums/ContentManagementFlagAction";
 
-export interface PowerpointContentFlags {
+export interface ExcelContentFlags {
+    DynamicDataExchange: ContentManagementFlagAction,
     EmbeddedFiles: ContentManagementFlagAction,
-    EmbeddedImages: ContentManagementFlagAction,
+    EmbeddedImages:ContentManagementFlagAction,
     ExternalHyperlinks: ContentManagementFlagAction,
     InternalHyperlinks: ContentManagementFlagAction,
     Macros: ContentManagementFlagAction,
