+++ server/src/common/models/PolicyManagementService/Policy/AdaptionPolicy/ContentFlags/PdfContentFlags.ts
@@ -1,4 +1,4 @@
-import { ContentManagementFlagAction } from "../../../../../../src/common/models/enums/ContentManagementFlagAction";
+import { ContentManagementFlagAction } from "../../../../enums/ContentManagementFlagAction";
 
 export interface PdfContentFlags {
     acroform: ContentManagementFlagAction,
