+++ server/src/common/models/PolicyManagementService/Policy/NcfsPolicy/NcfsOptions.ts
@@ -1,4 +1,4 @@
-import { NcfsOption } from "../../../../../src/common/models/enums/NcfsOption";
+import { NcfsOption } from "../../../enums/NcfsOption";
 
 export interface NcfsOptions {
     unProcessableFileTypes: NcfsOption,
