+++ server/src/common/services/IPolicyManagementService.ts
@@ -1,6 +1,6 @@
 import { Logger } from "winston";
 import { GetPolicyByIdRequest } from "../models/PolicyManagementService/GetPolicyById/GetPolicyByIdRequest";
-import { Policy } from "../../../frontend/src/types/Policy/Policy";
+import { Policy } from "../models/PolicyManagementService/Policy/Policy";
 
 export default interface IPolicyManagementService {
     logger: Logger,
